requirement.txt
torch
torchvision
transformers
onnx
onnxruntime
Pillow
numpy
scikit-learn